dashboard/requirements.txt
dash
plotly
pandas
requests
psycopg2-binary
scikit-learn